+++ generators/python/pydantic/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.10.0-rc0] - 2024-06-24
+
+- Upgrade: The generator now consumes IRV49.
+
 ## [0.9.1] - 2024-06-19
 
 - Internal: The generator now consumes IRv46.